Rename FastBuildRouter -> LightRouter.
