+++ robot-server/robot_server/commands/router.py
@@ -2,7 +2,7 @@
 from typing import Annotated, Final, List, Literal, Optional, cast
 
 from fastapi import Depends, Query, status
-from server_utils.fastapi_utils.fast_build_router import FastBuildRouter
+from server_utils.fastapi_utils.light_router import FastBuildRouter
 
 from opentrons.protocol_engine import CommandIntent
 from opentrons.protocol_engine.errors import CommandDoesNotExistError
