+++ robot-server/robot_server/deck_configuration/router.py
@@ -6,7 +6,7 @@ from typing import Annotated, Union
 
 import fastapi
 from starlette.status import HTTP_422_UNPROCESSABLE_ENTITY
-from server_utils.fastapi_utils.fast_build_router import FastBuildRouter
+from server_utils.fastapi_utils.light_router import FastBuildRouter
 
 from opentrons_shared_data.deck.types import DeckDefinitionV5
