+++ robot-server/robot_server/error_recovery/settings/router.py
@@ -4,7 +4,7 @@
 from typing import Annotated
 
 import fastapi
-from server_utils.fastapi_utils.fast_build_router import FastBuildRouter
+from server_utils.fastapi_utils.light_router import FastBuildRouter
 
 from robot_server.service.json_api import PydanticResponse, RequestModel, SimpleBody
 from .models import RequestData, ResponseData
