+++ robot-server/robot_server/maintenance_runs/router/base_router.py
@@ -10,7 +10,7 @@ from typing_extensions import Literal
 
 from fastapi import Depends, status
 from pydantic import BaseModel, Field
-from server_utils.fastapi_utils.fast_build_router import FastBuildRouter
+from server_utils.fastapi_utils.light_router import FastBuildRouter
 
 from robot_server.errors.error_responses import ErrorDetails, ErrorBody
 from robot_server.service.dependencies import get_current_time, get_unique_id
