+++ robot-server/robot_server/maintenance_runs/router/labware_router.py
@@ -3,7 +3,7 @@ from typing import Annotated
 import logging
 
 from fastapi import Depends, status
-from server_utils.fastapi_utils.fast_build_router import FastBuildRouter
+from server_utils.fastapi_utils.light_router import FastBuildRouter
 
 from opentrons.protocol_engine import LabwareOffsetCreate, LabwareOffset
 from opentrons.protocols.models import LabwareDefinition
