+++ robot-server/robot_server/robot/router.py
@@ -1,5 +1,5 @@
 """Router for /robot endpoints."""
-from server_utils.fastapi_utils.fast_build_router import FastBuildRouter
+from server_utils.fastapi_utils.light_router import FastBuildRouter
 
 from .control.router import control_router
