+++ robot-server/robot_server/runs/router/labware_router.py
@@ -9,7 +9,7 @@ from opentrons_shared_data.labware.labware_definition import (
     LabwareDefinition as SD_LabwareDefinition,
 )
 
-from server_utils.fastapi_utils.fast_build_router import FastBuildRouter
+from server_utils.fastapi_utils.light_router import FastBuildRouter
 
 from opentrons.protocol_engine import LabwareOffsetCreate, LabwareOffset
 from opentrons.protocols.models import LabwareDefinition
