+++ robot-server/robot_server/service/labware/router.py
@@ -7,7 +7,7 @@ from typing import Annotated, Optional
 from typing_extensions import Literal, NoReturn
 
 from fastapi import Depends, status
-from server_utils.fastapi_utils.fast_build_router import FastBuildRouter
+from server_utils.fastapi_utils.light_router import FastBuildRouter
 
 from opentrons_shared_data.errors import ErrorCodes
 from robot_server.errors.error_responses import ErrorDetails, ErrorBody
