+++ robot-server/robot_server/system/router.py
@@ -6,7 +6,7 @@ Endpoints include:
 """
 from datetime import datetime
 
-from server_utils.fastapi_utils.fast_build_router import FastBuildRouter
+from server_utils.fastapi_utils.light_router import LightRouter
 
 from robot_server.service.json_api.resource_links import ResourceLinkKey, ResourceLink
 
@@ -14,7 +14,7 @@ from .models import SystemTimeRequest, SystemTimeResponse, SystemTimeResponseAtt
 from .time_utils import get_system_time, set_system_time
 
 
-system_router = FastBuildRouter()
+system_router = LightRouter()
 """Router for /system endpoints."""
